Implement admin controller and make other changes

+++ app/Http/Controllers/Admin/AdminMovieController.php
@@ -12,14 +12,6 @@ use Illuminate\Http\Request;
 class AdminMovieController extends Controller
 {
     /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
      */
     public function store(MovieStoreRequest $request)
@@ -29,14 +21,6 @@ class AdminMovieController extends Controller
         $movie = Movie::create($validatedData);
 
         return new MovieResource($movie);
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
     }
 
     /**
@@ -60,6 +44,10 @@ class AdminMovieController extends Controller
     public function destroy(string $id)
     {
         $movie = Movie::find($id);
+
+        if (!$movie) {
+            return response()->json(['message' => 'Movie not found'], 404);
+        }
 
         $movie->delete();
